refactor: move ProductParameter to Product
